requirements.txt
anyio==3.6.1
brotlipy==0.7.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
cryptography==37.0.1
fastapi==0.81.0
#gunicorn==20.1.0
h11==0.13.0
idna==3.3
joblib==1.1.0
pip==22.2.2
pycparser==2.21
pydantic==1.10.1
pyOpenSSL==22.0.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.20.0
redbird==0.5.1
requests==2.28.1
rocketry==2.3.0
setuptools==65.3.0
six==1.16.0
sniffio==1.2.0
starlette==0.19.1
typing_extensions==4.3.0
urllib3==1.26.11
uvicorn==0.18.3
wheel==0.37.1
#win-inet-pton==1.1.0
aiofiles
python-multipart